This commit has everything from a working seeder for Postgres and Cassandra. Postgres is optimized and has functional GET request.
